backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
passlib[argon2]
python-jose[cryptography]